moved runner to execution package - removed shellext - added command split functionality using std lib strings
